# #1769 - Faster box approximation of VPolytope