Closes #1769.

In 2D this implementation is less efficient for a moderate number of vertices due to the additional conversion from `low`/`high` in the `Hyperrectangle` constructor. I guess in higher dimensions it will be more useful. Should I remove the dispatch for `VPolygon` or add an `invoke` of the old method based on the number of vertices?

```julia
julia> P1 = VPolytope([[0.0, 0.0], [2.0, 1.0], [1.0, 2.0]]);
julia> P2 = VPolytope([rand(2) for _ in 1:100]);
julia> P3 = VPolytope([rand(2) for _ in 1:1000]);

# this branch

julia> @time box_approximation(P1)
[0.0, 0.0][2.0, 2.0]
  0.000008 seconds (5 allocations: 416 bytes)
Hyperrectangle{Float64,Array{Float64,1},Array{Float64,1}}([1.0, 1.0], [1.0, 1.0])
julia> @time box_approximation(P2)
[0.001830275380493962, 0.005165832293896333][0.9791389830947983, 0.9914106144943142]
    0.000009 seconds (5 allocations: 416 bytes)
Hyperrectangle{Float64,Array{Float64,1},Array{Float64,1}}([0.49048462923764613, 0.4982882233941053], [0.48865435385715217, 0.49312239110020895])
julia> @time box_approximation(P3)
  0.000024 seconds (5 allocations: 416 bytes)
Hyperrectangle{Float64,Array{Float64,1},Array{Float64,1}}([0.4991620587806421, 0.4983966049474178], [0.49909969477169713, 0.4978833437247012])

# master

julia> @time box_approximation(P1)
  0.000004 seconds (3 allocations: 224 bytes)
Hyperrectangle{Float64,Array{Float64,1},Array{Float64,1}}([1.0, 1.0], [1.0, 1.0])
julia> @time box_approximation(P2)
  0.000008 seconds (3 allocations: 224 bytes)
Hyperrectangle{Float64,Array{Float64,1},Array{Float64,1}}([0.49048462923764613, 0.4982882233941053], [0.48865435385715217, 0.49312239110020895])
julia> @time box_approximation(P3)
  0.000049 seconds (3 allocations: 224 bytes)
Hyperrectangle{Float64,Array{Float64,1},Array{Float64,1}}([0.4991620587806421, 0.4983966049474178], [0.49909969477169713, 0.4978833437247012])
```