@mforets 2020-07-26
i had a look, see [this notebook](https://github.com/mforets/escritoire/blob/master/2020/Week31/VPolytope_BoxOA.ipynb). in low dims (usually <= 10) static arrays are much faster that regular arrays so it'd be nice that our code supports them. it is not the case with the new method but see the alternative implementation `ba2` in that notebook. it is also faster for dim 2 and different number of vertices. for dimension 10, `ba1` is faster, though i didn't make exhaustive checks!