@mforets 2020-07-26
> In 2D this implementation is less efficient for a moderate number of vertices

here `ba1` is faster than the version in master, while it allocates a bit more. see cells [15] and [9] in the notebook.