@schillic 2020-08-02
I added the ideas from your implementation with minor modifications.

Side note: I always get one more allocation than in your notebook (but even when I use your exact version, so this is not a problem of my modifications). Maybe you want to double-check...